Reorganize the OCR section
